Updated Time Pause & Player Movement
Time Pause now sets Time Scale to 0 instead of stopping all RigidBodies (makes it more universal, but movement is now in Update)
Player Movement is updated to use the character controller, so that they can move regardless of the Time Scale.
A restart function was also added, & the initial time pause delay is fixed to work now.